Removed isAdmin middleware

+++ app/Http/Controllers/AdminController.php
@@ -17,6 +17,9 @@ class AdminController extends Controller
 {
     // Renders the adminPanel at url/admin.
     function index() {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         $fishponds = Fishpond::all();
         return Inertia::render('AdminPanel', [
             'fishponds' => $fishponds
@@ -25,6 +28,9 @@ class AdminController extends Controller
 
     // Renders admin register page at url/admin/createNewAccount.
     function register() {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         return Inertia::render('Auth/AdminRegister', [
 
         ]);
@@ -32,6 +38,9 @@ class AdminController extends Controller
 
     // Creates new account recieved from Admin register page form through Post method and creates neccesary teams, wallet and subscription data.
     function createNewAccount(Request $request) {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         $input = $request->all();
         
         if (filter_var( $input['email'], FILTER_VALIDATE_EMAIL ) == false) {
@@ -79,6 +88,9 @@ class AdminController extends Controller
 
     // renders Admin edit existing accounts page while excluding the current admin to avoid accidental deletal
     function editExistingAccounts() {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         $users = User::where('id', '!=', auth::user()->id)->get()->load('Wallet')->load('Subscription');
         return Inertia::render('Auth/AdminOverview', [
             'users' => $users
@@ -87,6 +99,9 @@ class AdminController extends Controller
 
     // Deletes account if admin 
     function deleteAccount(Request $request) {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         $userId = $request->id;
         User::where('id', $userId)->delete();
         return redirect('/admin/editExistingAccounts');
@@ -94,6 +109,9 @@ class AdminController extends Controller
 
     // Renders edit fishpond overview
     function editFishpond(Request $request) {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         $fishpond = Fishpond::where('id', $request->id)->first();
         $fishes = Fish::all();
         return Inertia::render('AdminFishpondForm', [
@@ -104,6 +122,9 @@ class AdminController extends Controller
 
     // Edits the fishpond with data from Post form
     function confirmEditFishpond(Request $request) {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         $fishpond = Fishpond::where('id', $request->id)->first();
         $fishpond->name = $request->name;
         $fishpond->save();
@@ -112,6 +133,9 @@ class AdminController extends Controller
 
     // updates the fish type for a fishpond
     function confirmUpdateFishType(Request $request) {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         $fishpond = Fishpond::where('id', $request->id)->first();
         $fishpond->fish_id = $request->fishType;
         $fishpond->save();
@@ -120,6 +144,9 @@ class AdminController extends Controller
 
     // render transactions page for admin
     function showTransactions() {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         $transactions = Transaction::all()->load('user');
         return Inertia::render('AdminTransactions', [
             'transactions' => $transactions,
@@ -128,6 +155,9 @@ class AdminController extends Controller
 
     // renders the edit sensors page
     function editSensors(Request $request) {
+        if (auth::user()->admin != 1) {
+            return redirect('/dashboard');
+        }
         if (auth::user()->admin != 1) {
             return redirect('/dashboard');
         }
